Rename stabilizer_ui.widgets -> stabilizer_ui.ui
